requirements.txt
streamlit
torch
torchvision
pillow
numpy
matplotlib
fastapi
uvicorn
python-multipart